+++ hawkular-datamining-api/src/main/java/org/hawkular/datamining/api/util/Eager.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.hawkular.datamining.inventory.util;
+package org.hawkular.datamining.api.util;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
