+++ hawkular-datamining-api/src/main/resources/META-INF/services/javax.enterprise.inject.spi.Extension
@@ -1,0 +1,1 @@
+org.hawkular.datamining.api.util.EagerExtension
