--- hawkular-datamining-inventory-integration/src/main/resources/META-INF/services/javax.enterprise.inject.spi.Extension
@@ -1,1 +1,0 @@
-org.hawkular.datamining.inventory.util.EagerExtension
